Add clipping of lines

+++ points.cpp
@@ -1,22 +1,22 @@
 #include "points.h"
 
-Points::Points(int point_one, int point_two) {
+Points::Points(double point_one, double point_two) {
     this->point_one = point_one;
     this->point_two = point_two;
 }
 
-int Points::get_firstPoint() {
+double Points::get_firstPoint() {
     return this->point_one;
 }
 
-int Points::get_secondPoint() {
+double Points::get_secondPoint() {
     return this->point_two;
 }
 
-void Points::set_x(int newX) {
+void Points::set_x(double newX) {
     this->point_one = newX;
 }
 
-void Points::set_y(int newY) {
+void Points::set_y(double newY) {
     this->point_two = newY;
 }
